Convert Player EMC to being a BigInteger (#1928)
* Initial experimentation with converting player EMC to being a BigInteger

* Make EMC formatter obey the local locale

* Address the TODOs added

* Add remaining postfixes from https://en.wikipedia.org/wiki/Names_of_large_numbers as it doesn't hurt space wise to have 9 more entries in the lang file, and it makes it even more unlikely that someone will somehow reach a value that we can't represent

* Fix by mistake including 10^303 as if it was 10^66

* Fallback to using the too much emc translation key if the number would go past our max supported postfix

* Fix some rebase issues

+++ src/main/java/moze_intel/projecte/impl/TransmutationOffline.java
@@ -2,6 +2,7 @@ package moze_intel.projecte.impl;
 
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableList;
+import java.math.BigInteger;
 import moze_intel.projecte.PECore;
 import moze_intel.projecte.api.ProjectEAPI;
 import moze_intel.projecte.api.capabilities.IKnowledgeProvider;
@@ -123,12 +124,12 @@ public class TransmutationOffline
             }
 
             @Override
-            public long getEmc() {
+            public BigInteger getEmc() {
                 return toCopy.getEmc();
             }
 
             @Override
-            public void setEmc(long emc) {}
+            public void setEmc(BigInteger emc) {}
 
             @Override
             public void sync(@Nonnull ServerPlayerEntity player) {
